backend/requirements.txt
# Core dependencies for CrewAI implementation
crewai>=0.51.0
crewai-tools>=0.4.0

# LLM providers
anthropic>=0.25.0
openai>=1.30.0
langfuse>=3.0.0

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=13.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.41
pgvector>=0.3.0
alembic>=1.13.0

# Cache and Queue
redis>=5.0.0

# Vector Store
pinecone-client>=3.0.0

# Data validation
pydantic>=2.11.0
pydantic-settings>=2.3.0

# Utils
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=24.2.0
jinja2>=3.1.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.4.0